+++ satella/coding/structures/__init__.py
@@ -1,6 +1,13 @@
+from .dict_object import *
+from .immutable import Immutable
+from .singleton import Singleton
+from .structures import *
+from .typednamedtuple import *
+
 __all__ = [
     'typednamedtuple',
     'Heap',
+    'SetHeap',
     'TimeBasedHeap',
     'OmniHashableMixin',
     'Singleton',
@@ -8,9 +15,3 @@ __all__ = [
     'apply_dict_object',
     'Immutable'
 ]
-
-from .dict_object import *
-from .immutable import Immutable
-from .singleton import Singleton
-from .structures import *
-from .typednamedtuple import *